libeigen: remove obsolete setting for default out-of-tree build
Since commit 783fb88f@openembedded-core, out-of-tree builds of
cmake-based recipes are the default, and the variables
OECMAKE_BUILDPATH and OECMAKE_SOURCEPATH were deprecated.

With the commit 2c23d7ab@openembedded-core, the variable
OECMAKE_SOURCEPATH was put back into action, but this causes the
libeigen recipe to fail as OECMAKE_SOURCEPATH is set in the
recipe, but OECMAKE_BUILDPATH is set to the bbclass default.

This commit simply removes both variables from the libeigen recipe
and hence sets the recipe to do a default out-of-tree build.
This resolves the build failure of libeigen with
2c23d7ab@openembedded-core, and furthermore, the adjusted recipe
still works with the cmake.bbclass of 783fb88f@openembedded-core.

The build failure was discovered during the regression testing of
the meta-ros layer. The exact build error message and its
investigation is recorded in issue #276 of the meta-ros issue
tracker at https://github.com/bmwcarit/meta-ros/issues/276.

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@oss.bmw-carit.de>
Tested-by: Ash Charles <ashcharles@gmail.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>
Signed-off-by: Roman Meshkevich <roman.meshkevich@lanit-tercom.com>